# api/openapi/compliance-context.yaml
openapi: 3.0.3
info:
  title: Compliance Context Service API
  version: 1.0.0
  description: Compliance screening APIs.
servers:
  - url: https://api.sandbox.openfinance.ae
paths:
  /api/v1/compliance/screen:
    post:
      summary: Run compliance screening
      operationId: screenCompliance
      description: |-
        Screens once per transactionId. A retry for the same customer returns
        the stored result; the same transactionId for another customer is a 409.
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceScreeningRequest'
      responses:
        '201':
          description: Compliance result created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceScreeningResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Transaction already screened for another customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance/screenings/{transactionId}:
    get:
      summary: Find compliance result by transaction id
      operationId: findComplianceResult
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Compliance result found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceScreeningResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No screening result for this transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dpopAuth:
      type: http
      scheme: bearer
      bearerFormat: DPoP
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
    DPoP:
      name: DPoP
      in: header
      required: true
      schema:
        type: string
    InteractionId:
      name: X-FAPI-Interaction-ID
      in: header
      required: true
      schema:
        type: string
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
  schemas:
    ComplianceScreeningRequest:
      type: object
      required: [transactionId, customerId, amount]
      properties:
        transactionId:
          type: string
        customerId:
          type: string
        amount:
          type: number
          format: double
        sanctionsHit:
          type: boolean
        kycVerified:
          type: boolean
        pep:
          type: boolean
    ComplianceScreeningResponse:
      type: object
      required: [screeningId, transactionId, customerId, decision, reasons, checkedAt]
      properties:
        screeningId:
          type: string
        transactionId:
          type: string
        customerId:
          type: string
        decision:
          type: string
          enum: [PASS, REVIEW, FAIL]
        reasons:
          type: array
          items:
            type: string
        checkedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [code, message, interactionId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time